[ticket/9610] Make cache and store settings configurable (minor feature)
The "store" directory is set in the ACP, the "cache" directory is set via an optional constant (which would need to be defined in config.php or common.php).  They cannot both be set in the ACP because the cache object would then require the config array.  As it is now, getting the config array requires the cache object.

PHPBB3-9610